# main/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prospera",
    "description": "Prospera Market Simulator API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/portfolios/{environment_id}/cash/deposit": {
      "post": {
        "tags": [
          "portfolios"
        ],
        "summary": "Add Virtual Cash",
        "description": "Deposit virtual cash into environment.",
        "operationId": "add_virtual_cash_api_v1_portfolios__environment_id__cash_deposit_post",
        "parameters": [
          {
            "name": "environment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Environment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashAdjustmentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Add Virtual Cash Api V1 Portfolios  Environment Id  Cash Deposit Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{environment_id}/cash/withdraw": {
      "post": {
        "tags": [
          "portfolios"
        ],
        "summary": "Withdraw Virtual Cash",
        "description": "Withdraw virtual cash from environment.",
        "operationId": "withdraw_virtual_cash_api_v1_portfolios__environment_id__cash_withdraw_post",
        "parameters": [
          {
            "name": "environment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Environment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashAdjustmentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Withdraw Virtual Cash Api V1 Portfolios  Environment Id  Cash Withdraw Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{environment_id}/buy": {
      "post": {
        "tags": [
          "portfolios"
        ],
        "summary": "Buy Stock",
        "description": "Buy stock in environment.",
        "operationId": "buy_stock_api_v1_portfolios__environment_id__buy_post",
        "parameters": [
          {
            "name": "environment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Environment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TradeOrderInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Buy Stock Api V1 Portfolios  Environment Id  Buy Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{environment_id}/sell": {
      "post": {
        "tags": [
          "portfolios"
        ],
        "summary": "Sell Stock",
        "description": "Sell stock in environment.",
        "operationId": "sell_stock_api_v1_portfolios__environment_id__sell_post",
        "parameters": [
          {
            "name": "environment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Environment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TradeOrderInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Sell Stock Api V1 Portfolios  Environment Id  Sell Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{environment_id}/holdings": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "List Holdings",
        "description": "Get all holdings in environment.",
        "operationId": "list_holdings_api_v1_portfolios__environment_id__holdings_get",
        "parameters": [
          {
            "name": "environment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Environment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HoldingView"
                  },
                  "title": "Response List Holdings Api V1 Portfolios  Environment Id  Holdings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{environment_id}/transactions": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "List Transactions",
        "description": "Get transaction history for environment.",
        "operationId": "list_transactions_api_v1_portfolios__environment_id__transactions_get",
        "parameters": [
          {
            "name": "environment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Environment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionView"
                  },
                  "title": "Response List Transactions Api V1 Portfolios  Environment Id  Transactions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{environment_id}/performance": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "Get Performance",
        "description": "Get portfolio performance metrics.",
        "operationId": "get_performance_api_v1_portfolios__environment_id__performance_get",
        "parameters": [
          {
            "name": "environment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Environment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioPerformanceView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{environment_id}/sip": {
      "post": {
        "tags": [
          "portfolios"
        ],
        "summary": "Create Sip Plan",
        "description": "Create a recurring (SIP) investment plan for this portfolio.",
        "operationId": "create_sip_plan_api_v1_portfolios__environment_id__sip_post",
        "parameters": [
          {
            "name": "environment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Environment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSipPlanInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SipPlanView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "List Sip Plans",
        "description": "List the recurring (SIP) plans in this portfolio.",
        "operationId": "list_sip_plans_api_v1_portfolios__environment_id__sip_get",
        "parameters": [
          {
            "name": "environment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Environment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SipPlanView"
                  },
                  "title": "Response List Sip Plans Api V1 Portfolios  Environment Id  Sip Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{environment_id}/sip/{plan_id}": {
      "delete": {
        "tags": [
          "portfolios"
        ],
        "summary": "Cancel Sip Plan",
        "description": "Cancel a recurring (SIP) plan. Executed installments stay in history.",
        "operationId": "cancel_sip_plan_api_v1_portfolios__environment_id__sip__plan_id__delete",
        "parameters": [
          {
            "name": "environment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Environment Id"
            }
          },
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cancel Sip Plan Api V1 Portfolios  Environment Id  Sip  Plan Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CashAdjustmentInput": {
        "properties": {
          "environment_id": {
            "type": "string",
            "title": "Environment Id"
          },
          "amount": {
            "$ref": "#/components/schemas/Money"
          }
        },
        "type": "object",
        "required": [
          "environment_id",
          "amount"
        ],
        "title": "CashAdjustmentInput"
      },
...
      "CreateSipPlanInput": {
        "properties": {
          "environment_id": {
            "type": "string",
            "title": "Environment Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "amount": {
            "$ref": "#/components/schemas/Money"
          },
          "frequency": {
            "$ref": "#/components/schemas/SipFrequency",
            "default": "monthly"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "environment_id",
          "symbol",
          "amount"
        ],
        "title": "CreateSipPlanInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HoldingView": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "average_cost": {
            "type": "string",
            "title": "Average Cost"
          },
          "market_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Value"
          },
          "unrealized_pnl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unrealized Pnl"
          },
          "return_percentage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Percentage"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "quantity",
          "average_cost"
        ],
        "title": "HoldingView"
      },
...
      "Money": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "INR"
          }
        },
        "type": "object",
        "required": [
          "amount"
        ],
        "title": "Money"
      },
...
      "PortfolioPerformanceView": {
        "properties": {
          "environment_id": {
            "type": "string",
            "title": "Environment Id"
          },
          "cash_balance": {
            "type": "string",
            "title": "Cash Balance"
          },
          "invested_amount": {
            "type": "string",
            "title": "Invested Amount"
          },
          "portfolio_value": {
            "type": "string",
            "title": "Portfolio Value"
          },
          "unrealized_pnl": {
            "type": "string",
            "title": "Unrealized Pnl"
          },
          "return_percentage": {
            "type": "number",
            "title": "Return Percentage"
          }
        },
        "type": "object",
        "required": [
          "environment_id",
          "cash_balance",
          "invested_amount",
          "portfolio_value",
          "unrealized_pnl",
          "return_percentage"
        ],
        "title": "PortfolioPerformanceView"
      },
...
      "SipFrequency": {
        "type": "string",
        "enum": [
          "weekly",
          "monthly"
        ],
        "title": "SipFrequency",
        "description": "How often a systematic investment plan contributes."
      },
      "SipPlanView": {
        "properties": {
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "environment_id": {
            "type": "string",
            "title": "Environment Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "symbol_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol Name"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "frequency": {
            "type": "string",
            "title": "Frequency"
          },
          "day_of_month": {
            "type": "integer",
            "title": "Day Of Month"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "next_run_date": {
            "type": "string",
            "format": "date",
            "title": "Next Run Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "installments_run": {
            "type": "integer",
            "title": "Installments Run"
          },
          "installments_skipped": {
            "type": "integer",
            "title": "Installments Skipped"
          },
          "last_run_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run At"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "environment_id",
          "symbol",
          "symbol_name",
          "amount",
          "frequency",
          "day_of_month",
          "start_date",
          "next_run_date",
          "end_date",
          "status",
          "installments_run",
          "installments_skipped",
          "last_run_at",
          "created_at"
        ],
        "title": "SipPlanView"
      },
...
      "TradeOrderInput": {
        "properties": {
          "environment_id": {
            "type": "string",
            "title": "Environment Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "order_type": {
            "$ref": "#/components/schemas/TransactionType"
          }
        },
        "type": "object",
        "required": [
          "environment_id",
          "symbol",
          "quantity",
          "order_type"
        ],
        "title": "TradeOrderInput"
      },
      "TransactionType": {
        "type": "string",
        "enum": [
          "buy",
          "sell",
          "deposit",
          "withdrawal"
        ],
        "title": "TransactionType"
      },
      "TransactionView": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "transaction_type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "executed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Executed At"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "symbol",
          "transaction_type",
          "quantity",
          "amount",
          "executed_at"
        ],
        "title": "TransactionView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}